Retrait par sous-unité (comprimé) + fix foyer auth
- quantite passe en double pour supporter les fractions d'unités
- Bouton "Retirer 1 Comprimé" dans le détail quand quantiteParUnite défini
- Affichage "2 Plaquettes · 60 Comprimés" avec suivi précis
- Auth : guard anti-concurrence sur _onAuthChange, supprime double appel signIn/signUp

+++ lib/providers/medication_provider.dart
@@ -80,14 +80,14 @@ class MedicationProvider with ChangeNotifier {
     await load();
   }
 
-  Future<void> addStock(String? medicationId, int quantite) async {
+  Future<void> addStock(String? medicationId, double quantite) async {
     final fid = _auth?.currentFamilyId;
     if (fid == null || medicationId == null) return;
     await FirestoreRepository.addStock(fid, medicationId, quantite);
     await load();
   }
 
-  Future<void> takeStock(String? medicationId, int quantite) async {
+  Future<void> takeStock(String? medicationId, double quantite) async {
     final fid = _auth?.currentFamilyId;
     if (fid == null || medicationId == null) return;
     await FirestoreRepository.takeStock(fid, medicationId, quantite);
